file struct refactor + Readme for client part
