+++ client/src/pages/AuthPage/SubviewsAuthPage/RegistrationView/RegistrationView.tsx
