+++ client/src/pages/AuthPage/SubviewsAuthPage/RegistrationView/useRegistrationView.ts
